fix(maestro): surface re-exported component prop hovers
