+++ docs/release/typed-editor-oracle-matrix.md
@@ -14,6 +14,7 @@ surface honest without turning #4585 into an umbrella implementation PR.
 | Packaged VS Code host smoke rejects empty hover and `Ref<unknown>`/`ComputedRef<unknown>`/`MaybeRef<unknown>` for reactive values. | Covered | `editors/vscode/test/suite/real-scenario.cjs`                                                                                                                                                                                                              | #4589        |
 | Imported SFC hovers show props, emits, slots, model, and marker-free component contracts in LSP script surfaces.                   | Covered | `tests/tooling/lsp-imported-component-hover-type-backed.test.ts`; `tests/tooling/lsp-imported-component-reexport-hover-type-backed.test.ts`; CI: test-scripts runs tests/tooling with VIZE_TEST_REQUIRE_TSGO=1 (`tools/vite-plus/tasks/test-benchmark.ts`) | #4591        |
 | Component event hovers show child `defineEmits` payloads, and event definitions jump to the authored emit key.                     | Covered | `tests/tooling/lsp-component-event-hover-type-backed.test.ts`; CI: test-scripts runs tests/tooling with VIZE_TEST_REQUIRE_TSGO=1 (`tools/vite-plus/tasks/test-benchmark.ts`)                                                                               | #4592        |
+| Re-exported and package component prop hovers show child prop types instead of generic directive/null responses.                   | Covered | `tests/tooling/lsp-component-reexport-navigation.test.ts`; CI: test-scripts runs tests/tooling with VIZE_TEST_REQUIRE_TSGO=1 (`tools/vite-plus/tasks/test-benchmark.ts`)                                                                                   | #4592        |
 | Component `v-model` template navigation resolves to child `defineModel` declarations.                                              | Covered | `tests/tooling/lsp-component-v-model-type-backed.test.ts`; CI: test-scripts runs tests/tooling with VIZE_TEST_REQUIRE_TSGO=1 (`tools/vite-plus/tasks/test-benchmark.ts`)                                                                                   | #4592        |
 | Static `ref="..."` template ref values hover with precise `useTemplateRef()` types and jump to authored bindings.                  | Covered | `tests/tooling/lsp-reactive-hover-type-backed.test.ts`; CI: test-scripts runs tests/tooling with VIZE_TEST_REQUIRE_TSGO=1 (`tools/vite-plus/tasks/test-benchmark.ts`)                                                                                      | #4592        |
 | Non-VSCode editor-host smoke rejects reactive hover degradation.                                                                   | Covered | `editors/nvim/test/ref_surface_hover.lua`; `editors/nvim/test/vize_e2e_expected.lua`; CI: editor-host-smoke runs the packaged Neovim real-server scenario (`.github/actions/vscode-host-smoke/action.yml`)                                                 | #4589        |
