+++ tests/tooling/typed-editor-oracle-matrix.test.ts
@@ -4,247 +4,7 @@ import path from "node:path";
 import { test } from "node:test";
 
 import { root } from "./support/lsp/paths.ts";
-
-type MatrixRow = {
-  evidence: Evidence[];
-  followUp: string;
-  id: string;
-  status: "covered" | "known-gap";
-};
-
-type Evidence =
-  | { kind: "ci"; label: string; path: string; requiredText: string[] }
-  | { kind: "file"; path: string; requiredText: string[] }
-  | { kind: "pending-pr"; pr: number; reason: string };
-
-const rowsRequiringExecutedEvidence = new Set([
-  "lsp-imported-component-contract-hover",
-  "lsp-component-event-contract-navigation",
-  "lsp-component-v-model-navigation",
-  "lsp-static-template-ref-navigation",
-  "non-vscode-host-reactive-hover-surface",
-]);
-
-const toolingTestCiEvidence: Evidence = {
-  kind: "ci",
-  label: "test-scripts runs tests/tooling with VIZE_TEST_REQUIRE_TSGO=1",
-  path: "tools/vite-plus/tasks/test-benchmark.ts",
-  requiredText: [
-    "VIZE_TEST_REQUIRE_TSGO=1 node --test --test-concurrency=1 tests/tooling/*.test.ts",
-  ],
-};
-
-const matrix: MatrixRow[] = [
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "crates/vize/tests/check_text_diagnostics_cli.rs",
-        requiredText: [
-          "const count: string = 0;",
-          "Type 'number' is not assignable to type 'string'.",
-        ],
-      },
-    ],
-    followUp: "#4587",
-    id: "cli-authored-script-diagnostics",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "tests/tooling/lsp-authored-script-diagnostics.test.ts",
-        requiredText: ["const a: string = 1", "params.version === 2"],
-      },
-    ],
-    followUp: "#4587",
-    id: "lsp-authored-script-diagnostics",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "crates/vize_canon/src/sfc_typecheck/tests/fallthrough_ranges.rs",
-        requiredText: [
-          "fallthrough_diagnostic_range_uses_authored_template_offsets",
-          "diagnostic.start > script_start",
-        ],
-      },
-    ],
-    followUp: "#4586",
-    id: "cli-fallthrough-template-range",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "tests/tooling/lsp-fallthrough-attrs.test.ts",
-        requiredText: [
-          "publishes and clears fallthrough attribute diagnostics",
-          'id="outer"',
-          'diagnostic.code === "fallthrough-attrs"',
-          "$attrs.class",
-          "PlainFragment.vue",
-          "line: 5, character: 2",
-        ],
-      },
-    ],
-    followUp: "#4586",
-    id: "lsp-fallthrough-template-range",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "tests/tooling/lsp-hover-type-backed.test.ts",
-        requiredText: [
-          "hover and definition answer authored template anchors with backend type text",
-          "component slot hover must select the authored slot name",
-        ],
-      },
-    ],
-    followUp: "#4588 #4592",
-    id: "lsp-hover-definition-template-anchors",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "tests/snapshots/check/jsx-intrinsic-globals-oracle.ts",
-        requiredText: ["JSX.IntrinsicElements", "[TS7026]"],
-      },
-    ],
-    followUp: "#4590",
-    id: "cli-jsx-intrinsic-globals",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "editors/vscode/test/suite/real-scenario.cjs",
-        requiredText: ["refSurfaceHovers", "Ref<unknown>|ComputedRef<unknown>|MaybeRef<unknown>"],
-      },
-    ],
-    followUp: "#4589",
-    id: "vscode-host-reactive-hover-surface",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "tests/tooling/lsp-imported-component-hover-type-backed.test.ts",
-        requiredText: [
-          "script hover describes imported SFC contracts instead of generated markers",
-          "const Child: VueComponent",
-          "__vizeComponentMarker|__vizeRawProps|__VizeComponentConstructor",
-        ],
-      },
-      {
-        kind: "file",
-        path: "tests/tooling/lsp-imported-component-reexport-hover-type-backed.test.ts",
-        requiredText: [
-          "script hover describes re-exported and package SFC component contracts",
-          "Vue component: PackageChild.vue",
-          "__vizeComponentMarker|__vizeRawProps|__VizeComponentConstructor",
-        ],
-      },
-      toolingTestCiEvidence,
-    ],
-    followUp: "#4591",
-    id: "lsp-imported-component-contract-hover",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "tests/tooling/lsp-component-event-hover-type-backed.test.ts",
-        requiredText: [
-          "component event hovers and definitions use child emit contracts",
-          "Component event",
-          "[value: string]",
-        ],
-      },
-      toolingTestCiEvidence,
-    ],
-    followUp: "#4592",
-    id: "lsp-component-event-contract-navigation",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "tests/tooling/lsp-component-v-model-type-backed.test.ts",
-        requiredText: [
-          "component v-model hover and definition use the child model contract",
-          "definition must jump to the child defineModel declaration",
-        ],
-      },
-      toolingTestCiEvidence,
-    ],
-    followUp: "#4592",
-    id: "lsp-component-v-model-navigation",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "tests/tooling/lsp-reactive-hover-type-backed.test.ts",
-        requiredText: [
-          "static template ref value",
-          'ref="button"',
-          "useTemplateRef<HTMLButtonElement>",
-          "definition must jump to the authored useTemplateRef binding",
-        ],
-      },
-      toolingTestCiEvidence,
-    ],
-    followUp: "#4592",
-    id: "lsp-static-template-ref-navigation",
-    status: "covered",
-  },
-  {
-    evidence: [
-      {
-        kind: "file",
-        path: "editors/nvim/test/ref_surface_hover.lua",
-        requiredText: [
-          "degraded to an unknown reactive type",
-          "script ref hover",
-          "template template-ref hover",
-        ],
-      },
-      {
-        kind: "file",
-        path: "editors/nvim/test/vize_e2e_expected.lua",
-        requiredText: [
-          "ref_surface_hovers",
-          "const count: Ref<number, number>",
-          "const doubled: ComputedRef<number>",
-          "const button: HTMLButtonElement | null",
-        ],
-      },
-      {
-        kind: "ci",
-        label: "editor-host-smoke runs the packaged Neovim real-server scenario",
-        path: ".github/actions/vscode-host-smoke/action.yml",
-        requiredText: ["vp run --workspace-root test:nvim-extension:real-server"],
-      },
-    ],
-    followUp: "#4589",
-    id: "non-vscode-host-reactive-hover-surface",
-    status: "covered",
-  },
-];
+import { matrix, rowsRequiringExecutedEvidence } from "./support/typed-editor-oracle-matrix.ts";
 
 test("typed editor oracle matrix covers or explicitly tracks every P0 slice", () => {
   assert.deepEqual(
@@ -259,6 +19,7 @@ test("typed editor oracle matrix covers or explicitly tracks every P0 slice", ()
       "vscode-host-reactive-hover-surface",
       "lsp-imported-component-contract-hover",
       "lsp-component-event-contract-navigation",
+      "lsp-component-prop-reexport-hover",
       "lsp-component-v-model-navigation",
       "lsp-static-template-ref-navigation",
       "non-vscode-host-reactive-hover-surface",
